requirements.txt
typer[all]>=0.9.0
pandas>=2.0.0
pyyaml>=6.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
jinja2>=3.1.0
pdfplumber>=0.10.0
rich>=13.0.0

